refactor(test-runs): extract shared result-mutation guards
Move the result permission check and the flip-justification + required-field
gate helpers out of submit-result into lib/services/resultGuards.ts so the new
edit-result endpoint can enforce them identically. No behavior change;
submit-result's route tests still pass.
